+++ src/main/java/com/example/proyectoIntegrador/service/TurnoService.java
@@ -1,6 +1,12 @@
 package com.example.proyectoIntegrador.service;
 
+import com.example.proyectoIntegrador.exceptions.BadRequestException;
+import com.example.proyectoIntegrador.exceptions.ResourceNotFoundException;
+import com.example.proyectoIntegrador.model.Odontologo;
+import com.example.proyectoIntegrador.model.Paciente;
 import com.example.proyectoIntegrador.model.Turno;
+import com.example.proyectoIntegrador.repository.OdontologoRepository;
+import com.example.proyectoIntegrador.repository.PacienteRepository;
 import com.example.proyectoIntegrador.repository.TurnoRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -14,9 +20,21 @@ public class TurnoService {
 
     private final TurnoRepository turnoRepository;
 
+    private final PacienteRepository pacienteRepository;
 
-    public Turno save(Turno turno){
-        return turnoRepository.save(turno);
+    private final OdontologoRepository odontologoRepository;
+
+
+    public Turno save(Turno turno) throws BadRequestException {
+        Optional<Paciente> traerPaciente = pacienteRepository.findById(turno.getId());
+        Optional<Odontologo> traerOdontologo = odontologoRepository.findById(turno.getId());
+        if (traerPaciente.isPresent()&&traerOdontologo.isPresent())
+            turnoRepository.save(turno);
+        else {
+            throw new BadRequestException("No fue posible registrar el turno");
+        }
+
+        return turno;
     }
 
     public List<Turno> getAll() {
@@ -31,7 +49,12 @@ public class TurnoService {
         return turnoRepository.save(turno);
     }
 
-    public void delete(long id){
+    public void delete(long id) throws ResourceNotFoundException {
+        Optional<Turno> traerTurno = get(id);
+        if (traerTurno.isPresent())
         turnoRepository.deleteById(id);
+        else {
+            throw new ResourceNotFoundException("No existe el turno con el id: " +id+ "Ingrese un id valido");
+        }
     }
 }
